Refactor portfolio structure: remove unused loading and project pages; add new sections for Hero, About, Projects, and Contact; implement smooth scrolling and update navigation links.
