+++ src/app/globals.css
@@ -122,6 +122,10 @@
   }
 }
 
+html {
+  scroll-behavior: smooth;
+}
+
 .prose iframe {
   width: 100%;
   aspect-ratio: 16/9;
