+++ src/components/sections/AboutSection.tsx
@@ -4,14 +4,15 @@ import { UserIcon, EmailIcon, LocationIcon, CVIcon } from "@/components/icons";
 import { useTranslations } from "next-intl";
 import ParcoursEducatif from "@/components/ParcoursEducatif";
 
-export default function AboutPage() {
-
+export default function AboutSection() {
   const t = useTranslations();
 
   return (
-    <div className="min-h-screen">
-      <main className="py-10">
-        <h1 className="text-4xl font-bold text-center mb-8">{t('AboutCard.title')}</h1>
+    <section id="about" className="min-h-screen py-10">
+      <div className="container mx-auto max-w-7xl px-4">
+        <h2 className="text-4xl md:text-6xl font-bold text-center mb-8">
+          {t('AboutCard.title')}
+        </h2>
 
         <div className="flex flex-col md:flex-row gap-6 justify-center mx-auto">
           <div className="bg-card border border-border rounded-xl p-6 shadow-sm md:w-[450px] m-3">
@@ -55,10 +56,10 @@ export default function AboutPage() {
             </div>
           </div>
         </div>
+        
         <ParcoursProfessionnel />
         <ParcoursEducatif />
-
-      </main>
-    </div>
+      </div>
+    </section>
   );
 }
