+++ src/hooks/useScrollSpy.ts
@@ -1,0 +1,27 @@
+'use client';
+import { useEffect, useState } from 'react';
+
+export function useScrollSpy(sectionIds: string[], offset: number = 100) {
+  const [activeSection, setActiveSection] = useState<string>('');
+
+  useEffect(() => {
+    const handleScroll = () => {
+      const scrollPosition = window.scrollY + offset;
+
+      for (let i = sectionIds.length - 1; i >= 0; i--) {
+        const section = document.getElementById(sectionIds[i]);
+        if (section && section.offsetTop <= scrollPosition) {
+          setActiveSection(sectionIds[i]);
+          break;
+        }
+      }
+    };
+
+    handleScroll(); // Initial check
+    window.addEventListener('scroll', handleScroll);
+    
+    return () => window.removeEventListener('scroll', handleScroll);
+  }, [sectionIds, offset]);
+
+  return activeSection;
+}
